turn menu into an array of hashes, each dish is a hash which has a "name" and a "price".
also, order_spec stipulated `let(:takeaway) { Takeaway.new }` but then takeaway was never referred to.
it's a little more readable if we ` let(:order) { Order.new }` and refer to `order` rather than `subject`
you could argue that `subject` is nicer as it's really obvious that it's the object which is under test, but this is a matter of taste.